Struct hackrf::HackRF

pub struct HackRF {
    // some fields omitted
}

A structure for managaing a HackRF device.

This structure contains methods for configuring the various settings of the HackRF. It also provides a method for managing received samples as a blocking stream. Currently transmitting samples and non-blocking streams are not supported.

Implementation notes

This library is a lightweight wrapper around the basic HackRF FFI interface. In addition to the basic functionality it provides a blocking stream interface for receiving samples (See: RxStream).

Rust's type system is used to ensure that resources are not destroy or modified while the HackRF is receiving samples (although it is possible there are problems with the implementation.). To stop receiving samples it is sufficient to let the stream drop out of scope. For example:

let context = try!(hackrf::init());
let mut device = try!(HackRF::open(&context));

// Configure HackRF

{
    let mut rx_stream = try!(device.rx_stream());
    // Get samples from rx stream
}

// Re-configure HackRF

{
    let mut rx_stream = try!(device.rx_stream());
    // Get samples from rx stream
}

Methods

impl HackRF

fn open(context: &HackRFContext) -> HackRFResult<HackRF>

Attempt to open a connected HackRF device.

fn set_freq(&mut self, freq: u64) -> HackRFResult<()>

Sets the center frequency (in Hz) of the HackRF: 5-6800 MHz.

fn set_amp_enable(&mut self, amp_enable: bool) -> HackRFResult<()>

Sets amp gain to on/off.

Warning: Enabling this in the presence of high power signals could damage the HackRF.

fn set_antenna_enable(&mut self, antenna_enable: bool) -> HackRFResult<()>

Set antenna enable to on/off.

fn set_lna_gain(&mut self, gain: u32) -> HackRFResult<()>

Set LNA gain: 0-40 dB, 8dB steps. If the specified gain falls outside of the valid range, the gain will be clamped.

fn set_vga_gain(&mut self, gain: u32) -> HackRFResult<()>

Set VGA (IF) gain: 0-62 dB, with 2dB steps. If the specified gain falls outside of the valid range, the gain will be clamped.

fn set_samp_rate(&mut self, samp_rate: f64) -> HackRFResult<()>

Sets the sample rate in Hz: 2-20 MHz.

fn set_baseband_filter_bw(&mut self, filter_bw: u32) -> HackRFResult<()>

Sets the baseband filter bandwidth.

Note: This must be performed after setting the sample rate.

fn rx_stream(&mut self) -> HackRFResult<RxStream>

Start an RX stream.