Struct hackrf::RxStream

pub struct RxStream<'a> {
    // some fields omitted
}

A structure that manages receiving I/Q samples from the HackRF.

Implementation details

To obtain samples from the HackRF libhackrf requires that users register a callback which is called whenever a new block of samples avaliable from the underlying USB API. This interface, while powerful, is somewhat difficult to use. Instead of trying to expose this interface in a safe way, this library instead builds a blocking stream abstraction.

Internally, this library does the following:

First, it creates a set of shared buffers, with an associated Mutex and Condvar and registers a RX callback with libhackrf that has access to the shared buffers. Multiple shared buffers are used to reduce the chance that samples will need to be overwritten due to processing delays.

The registered callback is called by libhackrf whenever more samples are avaliable and simply copies that samples into the next free shared buffer, and signals that a new set of samples is avaliable using the Condvar.

On the consumer side, when one of the next_sample functions is called the function checks if any samples are avaliable in its local buffer, if not it will attempt to obtain new samples from the shared buffer. If there are new samples avaliable, the oldest shared buffer is copied into the local buffer and handled as normal. Otherwise the function will block until new new samples are avaliable using the Condvar. The local buffer is used to reduce the synchronization overhead of using a shared buffer.

Methods

impl<'a> RxStream<'a>

fn next_sample(&mut self) -> (f32, f32)

Return the next I/Q sample converted to a floating point number, using a lookup table.

fn next_sample_raw(&mut self) -> (u8, u8)

Return the next raw I/Q sample from the HackRF.